MVC+Hibernate

+++ src/main/java/web/dao/UserDaoImpl.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Repository;
 import web.model.User;
 
 import javax.persistence.EntityManager;
+import javax.persistence.EntityNotFoundException;
 import javax.persistence.PersistenceContext;
 import java.util.List;
 
@@ -13,7 +14,6 @@ public class UserDaoImpl implements UserDao {
     @PersistenceContext
     private EntityManager em;
 
-
     @Override
     public void save(User user) {
         em.persist(user);
@@ -21,17 +21,27 @@ public class UserDaoImpl implements UserDao {
 
     @Override
     public void update(User user) {
+        if (findById(user.getId()) == null) {
+            throw new EntityNotFoundException("User with ID " + user.getId() + " not found for update.");
+        }
         em.merge(user);
     }
 
     @Override
     public void delete(User user) {
+        if (findById(user.getId()) == null) {
+            throw new EntityNotFoundException("User with ID " + user.getId() + " not found for deletion.");
+        }
         em.remove(em.contains(user) ? user : em.merge(user));
     }
 
     @Override
     public User findById(int id) {
-        return em.find(User.class, id);
+        User user = em.find(User.class, id);
+        if (user == null) {
+            throw new EntityNotFoundException("User with ID " + id + " not found.");
+        }
+        return user;
     }
 
     @Override
